requirements.txt
#TODO: FIX
jupyterlab==3.0.0
torch==1.7.0
torchvision==0.8.1
click==7.1.2
scipy==1.2.1
seaborn==0.11.0
tensorboard==1.15.0
tqdm==4.45.0
scikit-learn==0.21.3
requests==2.25.0
Pillow==8.0.1
pandas==1.1.0
numba==0.50.1
numpy==1.19.5
matplotlib==3.3.3
biosppy==0.6.1
eli5==0.10.1
PyYAML==5.3.1




